# openapi/schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/physiotherapists": {
      "get": {
        "operationId": "UsersController_findAllPhysiotherapists",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "searchText",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfPhysiotherapistProfileDto",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PhysiotherapistProfileDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/physiotherapists/{id}": {
      "get": {
        "operationId": "UsersController_findPhysiotherapist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhysiotherapistProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/patients": {
      "get": {
        "operationId": "UsersController_findAllPatients",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "searchText",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfPatientProfileDto",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PatientProfileDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/patient/{id}": {
      "get": {
        "operationId": "UsersController_findPatient",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfPatientProfileDto",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PatientProfileDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user.",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfApplicationUserDto",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ApplicationUserDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a paginated list of users.",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a single user by its unique ID.",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user by its unique ID.",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user by its unique ID.",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "EasyMotion API",
    "description": "The EasyMotion API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "Enter your Bearer token"
      }
    },
    "schemas": {
...
      "ApplicationUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier (UUID)",
            "example": "b3bf4d18-8dd0-43a1-b1da-fd3f7b9553a1"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "description": "User middle name",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "description": "User phone number",
            "example": "+1234567890"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "User birth date",
            "example": "1990-01-01"
          },
          "role": {
            "type": "string",
            "description": "User role in the system",
            "enum": [
              "USER",
              "ADMIN",
              "PHYSIOTHERAPIST"
            ],
            "default": "USER"
          },
          "isEmailVerified": {
            "type": "boolean",
            "description": "Indicates if the user email has been verified",
            "default": false,
            "example": false
          },
          "lastLogin": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last login",
            "example": "2025-12-31T23:59:59.999Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of when the user was created",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last update",
            "example": "2025-01-10T10:00:00.000Z"
          },
          "picturePath": {
            "type": "string",
            "nullable": true,
            "description": "Profile picture path"
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "description": "Indicates if two-factor authentication is enabled",
            "default": false
          },
          "physiotherapist": {
            "nullable": true,
            "description": "Physiotherapist data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PhysiotherapistDto"
              }
            ]
          },
          "patient": {
            "nullable": true,
            "description": "Patient data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatientDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "isEmailVerified",
          "lastLogin",
          "createdAt",
          "updatedAt",
          "picturePath",
          "twoFactorEnabled",
          "physiotherapist",
          "patient"
        ]
      },
...
      "PhysiotherapistProfileDto": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "nullable": true,
            "description": "The bio of the physiotherapist",
            "example": "I am a physiotherapist with a passion for helping people"
          },
          "specialization": {
            "type": "string",
            "description": "The specialization of the physiotherapist",
            "example": "I am a physiotherapist with a passion for helping people"
          },
          "publicPhoneNumber": {
            "type": "string",
            "description": "The public phone number of the physiotherapist",
            "example": "1234567890"
          },
          "publicEmail": {
            "type": "string",
            "description": "The public email of the physiotherapist",
            "example": "physio@example.com"
          },
          "publicAddress": {
            "type": "string",
            "description": "The public address of the physiotherapist",
            "example": "123 Main St, Anytown, USA"
          },
          "website": {
            "type": "string",
            "description": "The website of the physiotherapist",
            "example": "https://www.physio.com"
          },
          "socialMediaLinks": {
            "description": "The social media links of the physiotherapist",
            "example": [
              "https://www.facebook.com/physio",
              "https://www.instagram.com/physio"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "description": "User middle name",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "lastLogin": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last login",
            "example": "2025-12-31T23:59:59.999Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of when the user was created",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "picturePath": {
            "type": "string",
            "nullable": true,
            "description": "Profile picture path"
          },
          "applicationUserId": {
            "type": "string"
          }
        },
        "required": [
          "bio",
          "specialization",
          "publicPhoneNumber",
          "publicEmail",
          "publicAddress",
          "website",
          "socialMediaLinks",
          "firstName",
          "lastName",
          "lastLogin",
          "createdAt",
          "picturePath",
          "applicationUserId"
        ]
      },
      "PatientProfileDto": {
        "type": "object",
        "properties": {
          "sex": {
            "type": "string",
            "nullable": true,
            "description": "Sex of the user",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "height": {
            "type": "number",
            "nullable": true,
            "description": "Height in cm"
          },
          "weight": {
            "type": "number",
            "nullable": true,
            "description": "Weight in kg"
          },
          "smoker": {
            "type": "boolean",
            "nullable": true,
            "description": "Smoker"
          },
          "alcoholUnits": {
            "type": "number",
            "nullable": true,
            "description": "Alcohol consumption (integer)"
          },
          "activityLevel": {
            "type": "string",
            "nullable": true,
            "description": "Activity level",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "mobilityLevel": {
            "type": "string",
            "nullable": true,
            "description": "Mobility level",
            "enum": [
              "LIMITED",
              "MODERATE",
              "FULL"
            ]
          },
          "restingHeartRate": {
            "type": "number",
            "nullable": true,
            "description": "Resting heart rate"
          },
          "bloodPressure": {
            "type": "string",
            "nullable": true,
            "description": "Blood pressure"
          },
          "profession": {
            "type": "string",
            "nullable": true,
            "description": "Profession"
          },
          "sport": {
            "type": "string",
            "nullable": true,
            "description": "Sport"
          },
          "sportFrequency": {
            "type": "number",
            "nullable": true,
            "description": "Sport frequency"
          },
          "medications": {
            "type": "string",
            "nullable": true,
            "description": "Medications"
          },
          "allergies": {
            "type": "string",
            "nullable": true,
            "description": "Allergies"
          },
          "otherPathologies": {
            "type": "string",
            "nullable": true,
            "description": "Other concomitant pathologies"
          },
          "painZone": {
            "type": "string",
            "nullable": true,
            "description": "Pain zone"
          },
          "painIntensity": {
            "type": "number",
            "nullable": true,
            "description": "Pain intensity"
          },
          "painFrequency": {
            "type": "string",
            "nullable": true,
            "description": "Pain frequency"
          },
          "painCharacteristics": {
            "type": "string",
            "nullable": true,
            "description": "Pain characteristics"
          },
          "painModifiers": {
            "type": "string",
            "nullable": true,
            "description": "Aspects that modify the intensity of pain"
          },
          "sleepHours": {
            "type": "number",
            "nullable": true,
            "description": "Sleep hours"
          },
          "perceivedStress": {
            "type": "number",
            "nullable": true,
            "description": "Perceived stress"
          },
          "lastMedicalCheckup": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Last medical checkup"
          },
          "personalGoals": {
            "type": "string",
            "nullable": true,
            "description": "Personal goals"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Notes"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "description": "User middle name",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "description": "User phone number",
            "example": "+1234567890"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "User birth date",
            "example": "1990-01-01"
          },
          "lastLogin": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last login",
            "example": "2025-12-31T23:59:59.999Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of when the user was created",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "picturePath": {
            "type": "string",
            "nullable": true,
            "description": "Profile picture path"
          },
          "applicationUserId": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "lastLogin",
          "createdAt",
          "picturePath",
          "applicationUserId"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the new user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "Plain-text password for the new user (will be hashed)",
            "example": "MyStrongP@ssw0rd!"
          },
          "isEmailVerified": {
            "type": "boolean",
            "description": "Flag indicating if the email address has been verified",
            "example": true
          },
          "firstName": {
            "type": "string",
            "description": "First name of the new user",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "description": "Middle name of the new user (optional)",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the new user",
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number of the new user (optional)",
            "example": "+123456789"
          },
          "birthDate": {
            "type": "string",
            "description": "Birth date in ISO format (YYYY-MM-DD) (optional)",
            "example": "1990-01-01"
          },
          "sex": {
            "type": "string",
            "description": "Sex of the new user (optional)",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "default": "MALE"
          },
          "role": {
            "type": "string",
            "description": "Role of the new user (defaults to USER if not provided)",
            "enum": [
              "USER",
              "ADMIN",
              "PHYSIOTHERAPIST"
            ],
            "default": "USER"
          }
        },
        "required": [
          "email",
          "password",
          "isEmailVerified",
          "firstName",
          "lastName"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}